requirements.txt
loop-python-client
pytest
python-can
